refactor(deps): all shared directory files moved to packages directory and establish as a separate project
