+++ packages/schemas/src/auth/auth.types.ts
@@ -1,4 +1,4 @@
 import type { z } from 'zod';
-import type { AuthSchema } from './auth.schema.js';
+import type { AuthSchema } from '../auth/auth.schema';
 
 export type AuthType = z.infer<typeof AuthSchema>;
