+++ packages/schemas/src/common/email.schema.ts
