+++ packages/schemas/src/common/password.schema.ts
