+++ packages/schemas/src/sessions/refresh.session.types.ts
@@ -1,4 +1,4 @@
 import type { z } from 'zod';
-import type { RefreshSessionSchema } from './refresh.session.schema.js';
+import type { RefreshSessionSchema } from '../sessions/refresh.session.schema';
 
 export type RefreshSessionType = z.infer<typeof RefreshSessionSchema>;
