+++ packages/schemas/src/user/server/create.user.req.schema.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { CreateUserServerSchema } from './user.create.server.schema.js';
+import { CreateUserServerSchema } from './user.create.server.schema';
 
 export const CreateUserServerReqSchema = z.object({
 	body: z.object({
