+++ packages/schemas/src/user/server/update.user.req.schema.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { UpdateUserServerSchema } from './user.update.server.schema.js';
+import { UpdateUserServerSchema } from './user.update.server.schema';
 
 export const UpdateUserServerReqSchema = z.object({
 	body: z.object({
